requirements.txt
pytest==7.0.0
pytest-bdd==5.0.0
appium-python-client==4.2.1 
selenium==4.12.0
browserstack-local==1.2.9
